add full tests for Cezar
